Split into two classes
HamlLine holds the bulk of the HAML-parsing "logic". HamlParser
provides an API more befitting a parser -- render(), static functions,
etc.

A couple of instanceof checks had to change to be HamlLine instead of
HamlParser. render() also becomes a bit simpler; it no longer has to
determine if this instance is a single line or not, because it's
always a full file, and to parse a single line you call parseLine().